main/requirements/base.txt
Django==3.1.6
Pillow==8.1.2
PyMySQL==1.0.2
cryptography==3.3.1
reportlab==3.5.63
requests==2.25.1
qrcode==6.1
